+++ changelog.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - 2020-11-30
+### Added
+- Add priority fro events.
+- All the events will be sorted by priority.
+- Finish only need first 4 digits now.
+
 ## [0.1.1] - 2020-11-24
 ### Added
 - Timezone output will be local timezone.